+++ APP/components/math.js
@@ -1,15 +1,14 @@
 // all nameable, and default inline
 // numberable quoted block, possibly containing a proof: thm, lemma, prop
 // proof is automatically unquoted if inside thm block
-// the style of the word "proof" can be set to be less bold; this shouldn't be done manually, automatically done for proofs inside of thm blocks
+// the style of the word "proof" can be set to be not bold; this shouldn't be done manually, automatically done for proofs inside of thm blocks
 // default unquoted: defn, example; everything else default quoted
-// retypeable: thm, proof (e.g. to claim, justification); if renamed then not bold or underlined
+// retypeable: thm, proof (e.g. to claim, justification, corollary)
 // example can be plural, in which case it is set to outofline
 
-function render_statement(statement, outofline, type, name, num, unbolded ) {
+function render_statement(statement, outofline, type, name, num, title_style ) {
     let displayed_num = num != -1 ? <> {num}</> : <></>
     let displayed_name = name ? <>({name})</> : <></>
-    let title_style = 'italic text-highlight-subtle' + (unbolded ? '' : ' font-bold')
     let title = (
         <div className='inline-flex flex-wrap gap-1 text-left'>
             <span className={title_style}>{type}{displayed_num}.</span>
@@ -48,7 +47,7 @@ function render_statement(statement, outofline, type, name, num, unbolded ) {
     )
 }
 
-function render_thm(children, outofline, type, name, num) {
+function render_thm(children, outofline, type, name, num, title_style) {
     let statement = children
     let proof = null
     if(children.constructor===Array && children[children.length-1].type.name === 'Proof') {
@@ -57,7 +56,7 @@ function render_thm(children, outofline, type, name, num) {
     }
     return (
         <>
-            {render_statement(statement, outofline, type, name, num)}
+            {render_statement(statement, outofline, type, name, num, title_style)}
             {proof ? (outofline ? <div className="mt-2">{proof}</div> : proof) : <></>}
         </>
     )
@@ -72,20 +71,26 @@ function quote(react, unquoted, lines=false) {
 }
 
 export function Proof({ children, outofline=false, type='Proof', name='', unquoted=false, unbolded=false }) {
-    return quote(render_statement(children, outofline, type, name, -1, unbolded), unquoted)
+    let title_style = 'italic text-highlight-subtle' + (unbolded ? '' : ' font-bold')
+    return quote(render_statement(children, outofline, type, name, -1, title_style), unquoted)
 }
 export function Thm({ children, outofline=false, type='Theorem', name='', unquoted=false, num=-1 }) {
-    return quote(render_thm(children, outofline, type, name, num), unquoted, true, false) // if unquoted, then add lines instead
+    let title_style = 'italic text-highlight-strong font-bold'
+    return quote(render_thm(children, outofline, type, name, num, title_style), unquoted, true, false) // if unquoted, then add lines instead
 }
 export function Lemma({ children, outofline=false, name='', unquoted=false, num=-1 }) {
-    return quote(render_thm(children, outofline, 'Lemma', name, num), unquoted, false, false) // don't add lines
+    let title_style = 'italic text-highlight-strong font-bold'
+    return quote(render_thm(children, outofline, 'Lemma', name, num, title_style), unquoted, false, false) // don't add lines
 }
 export function Prop({ children, outofline=false, name='', unquoted=false, num=-1 }) {
-    return quote(render_thm(children, outofline, 'Prop', name, num), unquoted, false, false) // don't add lines
+    let title_style = 'italic text-highlight-strong font-bold'
+    return quote(render_thm(children, outofline, 'Prop', name, num, title_style), unquoted, false, false) // don't add lines
 }
 export function Defn({ children, outofline=false, name='', quoted=false}) {
-    return quote(render_statement(children, outofline, 'Defn', name, -1), !quoted, false, false) // don't add lines
+    let title_style = 'italic text-highlight-strong font-bold'
+    return quote(render_statement(children, outofline, 'Defn', name, -1, title_style), !quoted, false, false) // don't add lines
 }
 export function Example({ children, outofline=false, name='', quoted=false, plural=false }) {
-    return quote(render_statement(children, outofline||plural, plural ? 'Examples' : 'Example', name, -1), !quoted, false, false) // don't add lines
+    let title_style = 'italic text-highlight-subtle'
+    return quote(render_statement(children, outofline||plural, plural ? 'Examples' : 'Example', name, -1, title_style), !quoted, false, false) // don't add lines
 }