+++ APP/pages/maths/index.js
@@ -215,7 +215,7 @@ export default function Maths () {
                 <Link href="/maths/Analysis-Topology">
                     <div className="flex items-center space-x-1 text-lg underline">
                         
-                            <FaBook className="shrink-0"/>
+                            <FaBook className="shrink-0 text-course-icon"/>
                         
                         <p className="relative bottom-[1px]">Analysis-Topology</p>
                     </div>
@@ -227,7 +227,7 @@ export default function Maths () {
                 <Link href="/maths/proving-FTA">
                     <div className="flex items-center space-x-1 text-lg underline">
                         
-                            <AiFillFolder className="shrink-0"/>
+                            <AiFillFolder className="shrink-0 text-folder-icon"/>
                         
                         <p className="relative bottom-[1px]">proving-FTA</p>
                     </div>
@@ -239,7 +239,7 @@ export default function Maths () {
                 <Link href="/maths/research">
                     <div className="flex items-center space-x-1 text-lg underline">
                         
-                            <AiFillFolder className="shrink-0"/>
+                            <AiFillFolder className="shrink-0 text-folder-icon"/>
                         
                         <p className="relative bottom-[1px]">research</p>
                     </div>
@@ -251,7 +251,7 @@ export default function Maths () {
                 <Link href="/maths/olympiad">
                     <div className="flex items-center space-x-1 text-lg underline">
                         
-                            <AiFillFolder className="shrink-0"/>
+                            <AiFillFolder className="shrink-0 text-folder-icon"/>
                         
                         <p className="relative bottom-[1px]">olympiad</p>
                     </div>
@@ -266,7 +266,7 @@ export default function Maths () {
                 <Link href="/maths/Analysis-Topology">
                     <div className="flex items-center space-x-1 text-lg underline">
                         
-                            <FaBook className="shrink-0"/>
+                            <FaBook className="shrink-0 text-course-icon"/>
                         
                         <p className="relative bottom-[1px]">Analysis-Topology</p>
                     </div>
@@ -278,7 +278,7 @@ export default function Maths () {
                 <Link href="/maths/olympiad">
                     <div className="flex items-center space-x-1 text-lg underline">
                         
-                            <AiFillFolder className="shrink-0"/>
+                            <AiFillFolder className="shrink-0 text-folder-icon"/>
                         
                         <p className="relative bottom-[1px]">olympiad</p>
                     </div>
@@ -290,7 +290,7 @@ export default function Maths () {
                 <Link href="/maths/proving-FTA">
                     <div className="flex items-center space-x-1 text-lg underline">
                         
-                            <AiFillFolder className="shrink-0"/>
+                            <AiFillFolder className="shrink-0 text-folder-icon"/>
                         
                         <p className="relative bottom-[1px]">proving-FTA</p>
                     </div>
@@ -302,7 +302,7 @@ export default function Maths () {
                 <Link href="/maths/research">
                     <div className="flex items-center space-x-1 text-lg underline">
                         
-                            <AiFillFolder className="shrink-0"/>
+                            <AiFillFolder className="shrink-0 text-folder-icon"/>
                         
                         <p className="relative bottom-[1px]">research</p>
                     </div>