+++ templates/_app.js
@@ -1,11 +1,12 @@
 import '../styles/globals.css'
 import '../styles/codehilite.css'
-import { Fira_Code } from 'next/font/google'
-import localFont from 'next/font/local'
+import { Lato, Fira_Code } from 'next/font/google'
 import Head from 'next/head'
 
-const iAWriterQuattro = localFont({
-  src: '../fonts/iAWriterQuattroV.ttf'
+const lato = Lato({
+  weight: ['400', '700'],
+  subsets: ['latin'],
+  display: 'swap',
 })
 
 const fira = Fira_Code({
@@ -17,11 +18,11 @@ export default function App({ Component, pageProps }) {
   return (
     <>
       <Head>
-        <meta name="theme-color" content="#f4f4f0"/>
+        <meta name="theme-color" content="#ffffff"/>
       </Head>
       <style jsx global>{`
         :root {
-          --font-text: ${iAWriterQuattro.style.fontFamily};
+          --font-text: ${lato.style.fontFamily};
           --font-code: ${fira.style.fontFamily};
         }
       `}</style>